feat: fully implement main account

+++ db/mutations/accounts.ts
@@ -1,10 +1,33 @@
-import { eq } from "drizzle-orm";
+import { eq, or } from "drizzle-orm";
 import { db } from "../client";
-import { accountsTable, mainAccountsTable, SchemaAccount } from "../schema";
+import { getMainAccount } from "../queries/accounts";
+import { accountsTable, mainAccountsTable, SchemaAccount, transactionsTable } from "../schema";
+import { batchDeleteTransactions } from "./transactions";
 
-// todo: what happens to main account and related transactions
+// todo: fix react query undefined issue
 export const deleteAccount = async (id: SchemaAccount["id"]) => {
-  await db.delete(accountsTable).where(eq(accountsTable.id, id));
+  const mainAccount = await getMainAccount();
+  if (mainAccount?.account_id === id) {
+    // prevent the main account from being deleted
+    // TODO: log or throw error
+    return;
+  }
+
+  await db.transaction(async (tx) => {
+    // delete related transfer transactions
+    const transferTransactions = await db
+      .select()
+      .from(transactionsTable)
+      .where(
+        or(eq(transactionsTable.to_account_id, id), eq(transactionsTable.from_account_id, id))
+      );
+
+    // using this method to delete transfer transactions since it updates the balance of the
+    // other account in the transaction
+    // other transactions will be delete by delete cascade once the account is deleted
+    await batchDeleteTransactions(transferTransactions);
+    await db.delete(accountsTable).where(eq(accountsTable.id, id));
+  });
 };
 
 export const insertMainAccount = async (id: SchemaAccount["id"]) => {
